Implementando carrinho

+++ SGVE/SGVE-web/Models/CartHeaderViewModel.cs
@@ -5,6 +5,6 @@ namespace SGVE_web.Models
     {
         public long Id { get; set; }
         public string UserId { get; set; }
-        public double PurchaseAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
     }
 }
